+++ examples/flask-webhooks-example/instance/example_config.py
@@ -1,0 +1,11 @@
+# App Configs
+SECRET_KEY = 'P\tK\xa6\x99"\xadR)\xd2Z\xa7\xfb\x98IS'
+SQLALCHEMY_DATABASE_URI = 'sqlite:////tmp/test.db'
+SQLALCHEMY_TRACK_MODIFICATIONS = False
+NGROK_BASE_URL = "https://abcdefg.ngrok.io"
+
+# LaunchKey Configs
+ORGANIZATION_ID = "28a8c483-f1a2-4015-adbb-a31e6ea18b90"
+ORGANIZATION_KEY = open("/path/to/private.key").read()
+DIRECTORY_ID = "e243bf77-9966-4fbd-90ed-bb7f4405d9cf"
+SERVICE_ID = "babdb1b5-cb52-4231-9e72-c72732a47c00"